mcp-use-grupify/requirements.txt
langchain
langchain-openai
python-dotenv
openai>=1.0.0
fastembed
mysql-connector-python
streamlit
google-generativeai
langchain-google-genai
